Documentation, reorganization, API nation
A player changing worlds now boots anyone lacking cross-world permissions from their inventories.
Permissions is now the enum it so desperately was trying to become. The check method now is inside it.
Added a lot of javadocs, hoping to polish and actually release a semi-permanent API, similar to before the dark ages of single-version OpenInv.
Removing all the deprecated methods and some that aren't deprecated; getting ready to throw this whole mass onto the master branch, and no matter how much I dressed it up it'd break anyone depending on us anyway.
The checks using InventoryAccess are now marginally less efficient for our particular use case, but the slowdown does allow us a much more user-friendly API. It may even be worth changing IInventoryAccess to an entirely reflection-based system; it hasn't changed in the slightest since 1.4